requirements.txt
maafw==4.4.1
loguru
pillow
pytz